404 http code responses in TransactionController

+++ src/main/java/pl/myCompany/transaction/TransactionController.java
@@ -1,6 +1,10 @@
 package pl.myCompany.transaction;
 
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
 import java.util.Optional;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.ResponseEntity.BodyBuilder;
@@ -12,6 +16,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+@Slf4j
 @RestController
 public class TransactionController {
 
@@ -23,12 +28,26 @@ public class TransactionController {
 
   @GetMapping("/transactions")
   public ResponseEntity<Iterable<Transaction>> getTransactions() {
-    return ResponseEntity.ok(transactionService.getTransactions());
+    Iterable<Transaction> transactionsIterable = transactionService.getTransactions();
+    Iterator<Transaction> transactionIterator = transactionsIterable.iterator();
+    List<Transaction> transactionList = convertIterableToCollectionList(transactionIterator);
+    return ResponseEntity.ok(transactionList);
+  }
+
+  private List<Transaction> convertIterableToCollectionList(Iterator<Transaction> transactions) {
+    List<Transaction> transactionList = new ArrayList<>();
+    transactions.forEachRemaining(transactionList::add);
+    return transactionList;
   }
 
   @GetMapping("/transaction/{id}")
   public ResponseEntity<Optional<Transaction>> getTranscation(@PathVariable long id) {
-    return ResponseEntity.ok(transactionService.getTransaction(id));
+    Optional<Transaction> transaction = transactionService.getTransaction(id);
+    if (!transaction.isPresent()) {
+      log.info("Transaction with id = {} was not found", id);
+      return ResponseEntity.notFound().build();
+    }
+    return ResponseEntity.ok(transaction);
   }
 
   @PostMapping("/transaction")
@@ -42,7 +61,7 @@ public class TransactionController {
         description(transactionRequest.getDescription())
         .categoryId(transactionRequest.getCategoryId())
         .localDate(transactionRequest.getLocalDate())
-        .isPlanned(transactionRequest.isPlanned()).build(); //TODO isPlanned works?
+        .isPlanned(transactionRequest.isPlanned()).build(); //TODO Question: isPlanned does not works, why there is no getter for that boolean type parameter? The name of the field in get method is "planned", why?
     return convertedTransaction;
   }
 
@@ -53,13 +72,18 @@ public class TransactionController {
         .categoryId(transactionRequest.getCategoryId())
         .localDate(transactionRequest.getLocalDate())
         .isPlanned(transactionRequest.isPlanned())
-        .build(); //TODO isPlanned works?
+        .build(); //TODO same as 64
     transactionService.updateTransaction(id, newTransaction);
     return ResponseEntity.ok().build();
   }
 
   @DeleteMapping("/transaction/{id}")
   public ResponseEntity<?> deleteTransaction(@PathVariable long id) {
+    Optional<Transaction> transaction = transactionService.getTransaction(id);
+    if (!transaction.isPresent()) {
+      log.info("No able to delete transaction with id = {} because it was not found", id);
+      return ResponseEntity.notFound().build();
+    }
     transactionService.deleteTransaction(id);
     return ResponseEntity.ok().build();
   }
